+++ packages/browser-wallet-api/src/wallet-api.ts
@@ -35,8 +35,9 @@ class WalletApi extends EventEmitter implements IWalletApi {
     /**
      * Sends a sign request to the Concordium Wallet and awaits the users action
      */
-    public async signMessage(message: string): Promise<AccountTransactionSignature> {
+    public async signMessage(accountAddress: string, message: string): Promise<AccountTransactionSignature> {
         const response = await this.sendMessage<AccountTransactionSignature | undefined>(MessageType.SignMessage, {
+            accountAddress,
             message,
         });
 
@@ -53,7 +54,8 @@ class WalletApi extends EventEmitter implements IWalletApi {
     public async connect(): Promise<string | undefined> {
         const response = await this.messageHandler.sendMessage<string | undefined | false>(MessageType.Connect);
 
-        if (response === false) {
+        // TODO Response becomes === null when we would expect it to be undefined. Catching it here is a temporary quick-fix.
+        if (response === false || response === null) {
             throw new Error('Connection rejected');
         }
 
@@ -66,12 +68,14 @@ class WalletApi extends EventEmitter implements IWalletApi {
      * Sends a transaction to the Concordium Wallet and awaits the users action
      */
     public async sendTransaction(
+        accountAddress: string,
         type: AccountTransactionType,
         payload: AccountTransactionPayload,
         parameters?: Record<string, unknown>,
         schema?: string
     ): Promise<string> {
         const response = await this.sendMessage<string | undefined>(MessageType.SendTransaction, {
+            accountAddress,
             type,
             payload: stringify(payload),
             parameters,