+++ packages/browser-wallet/src/background/index.ts
@@ -5,7 +5,7 @@ import {
     MessageType,
 } from '@concordium/browser-wallet-message-hub';
 import { HttpProvider } from '@concordium/web-sdk';
-import { storedSelectedAccount, storedUrlWhitelist, storedJsonRpcUrl } from '@shared/storage/access';
+import { storedConnectedSites, storedSelectedAccount, storedJsonRpcUrl } from '@shared/storage/access';
 
 import JSONBig from 'json-bigint';
 import bgMessageHandler from './message-handler';
@@ -52,12 +52,18 @@ bgMessageHandler.handleMessage(createMessageTypeFilter(MessageType.JsonRpcReques
 });
 
 /**
- * Run condition which looks up url in stored whitelist. Runs handler if url is included in whitelist.
+ * Run condition which looks up URL in connected sites for the provided account. Runs handler if URL is included in connected sites.
  */
-const runIfWhitelisted: RunCondition<false> = async (_msg, sender) => {
-    const whitelist = (await storedUrlWhitelist.get()) ?? [];
+const runIfWhitelisted: RunCondition<false> = async (msg, sender) => {
+    const { accountAddress } = msg.payload;
+    const connectedSites = await storedConnectedSites.get();
 
-    if (sender.url !== undefined && whitelist.includes(sender.url)) {
+    if (!accountAddress || connectedSites === undefined) {
+        return { run: false, response: false };
+    }
+
+    const accountConnectedSites = connectedSites[accountAddress] ?? [];
+    if (sender.url !== undefined && accountConnectedSites.includes(sender.url)) {
         return { run: true };
     }
 
@@ -65,20 +71,62 @@ const runIfWhitelisted: RunCondition<false> = async (_msg, sender) => {
 };
 
 /**
- * Run condition which looks up url in stored whitelist. Runs handler if url is NOT included in whitelist.
+ * Finds the most prioritized account that is connected to the provided site.
+ * The priority is defined as:
+ * 1. If the selected account is connected, then that is returned.
+ * 1. The first account other than the selected account that is connected to the site. The order here
+ * is defined by the order of the entries of the stored connected sites.
+ * @param url the site to find an account that is connected to
+ * @returns the highest priority account address that is connected to the site with the provided URL.
+ */
+async function findPrioritizedAccountConnectedToSite(url: string): Promise<string | undefined> {
+    const selectedAccount = await storedSelectedAccount.get();
+    const connectedSites = await storedConnectedSites.get();
+
+    if (!selectedAccount || !connectedSites) {
+        return undefined;
+    }
+
+    const selectedAccountConnectedSites = connectedSites[selectedAccount] ?? [];
+    if (selectedAccountConnectedSites.includes(url)) {
+        return selectedAccount;
+    }
+
+    const connectedAccount = Object.entries(connectedSites).find((item) => item[1] && item[1].includes(url));
+    if (connectedAccount) {
+        return connectedAccount[0];
+    }
+
+    return undefined;
+}
+
+/**
+ * Run condition that runs the handler if the wallet is non-empty (an account exists), and no
+ * account in the wallet is connected to the sender URL.
+ *
+ * 1. If no selected account exists (the wallet is empty), then do not run and return undefined.
+ * 1. Else if the selected account is connected to the sender URL, then do not run and return the selected account address.
+ * 1. Else if any other account is connected to the sender URL, then do not run and return that account's address.
+ * 1. Else run the handler.
  */
 const runIfNotWhitelisted: RunCondition<string | undefined> = async (_msg, sender) => {
-    const whitelist = (await storedUrlWhitelist.get()) ?? [];
-
     if (!sender.url) {
-        throw new Error('Expected url to be available for sender.');
+        throw new Error('Expected URL to be available for sender.');
     }
 
-    if (whitelist.includes(sender.url)) {
-        const selectedAccount = await storedSelectedAccount.get();
-        return { run: false, response: selectedAccount };
+    const selectedAccount = await storedSelectedAccount.get();
+
+    // No accounts in the wallet.
+    if (selectedAccount === undefined) {
+        return { run: false, response: undefined };
     }
 
+    const accountConnectedToSite = await findPrioritizedAccountConnectedToSite(sender.url);
+    if (accountConnectedToSite) {
+        return { run: false, response: accountConnectedToSite };
+    }
+
+    // No account in the wallet is connected to the URL, so run the handler.
     return { run: true };
 };
 
@@ -97,10 +145,7 @@ const handleConnectionResponse: HandleResponse<string | undefined | false> = asy
     }
 
     if (response !== false) {
-        const whitelist = (await storedUrlWhitelist.get()) ?? [];
-        storedUrlWhitelist.set([...whitelist, sender.url]);
-
-        return storedSelectedAccount.get();
+        return findPrioritizedAccountConnectedToSite(sender.url);
     }
 
     return response;