+++ packages/browser-wallet/src/popup/shared/message-handler.ts
@@ -1,4 +1,4 @@
 import { ExtensionsMessageHandler } from '@concordium/browser-wallet-message-hub';
-import { storedUrlWhitelist } from '@shared/storage/access';
+import { storedConnectedSites, storedSelectedAccount } from '@shared/storage/access';
 
-export const popupMessageHandler = new ExtensionsMessageHandler(storedUrlWhitelist);
+export const popupMessageHandler = new ExtensionsMessageHandler(storedConnectedSites, storedSelectedAccount);