+++ packages/browser-wallet/src/popup/store/utils.ts
@@ -1,19 +1,19 @@
 import {
     StorageAccessor,
+    storedConnectedSites,
     storedCredentials,
     storedJsonRpcUrl,
     storedSelectedAccount,
     storedTheme,
-    storedUrlWhitelist,
 } from '@shared/storage/access';
 import { ChromeStorageKey } from '@shared/storage/types';
 import { atom, WritableAtom } from 'jotai';
 
 const accessorMap = {
+    [ChromeStorageKey.ConnectedSites]: storedConnectedSites,
     [ChromeStorageKey.Credentials]: storedCredentials,
     [ChromeStorageKey.SelectedAccount]: storedSelectedAccount,
     [ChromeStorageKey.JsonRpcUrl]: storedJsonRpcUrl,
-    [ChromeStorageKey.UrlWhitelist]: storedUrlWhitelist,
     [ChromeStorageKey.Theme]: storedTheme,
 };