+++ packages/browser-wallet-api-helpers/CHANGELOG.md
@@ -6,6 +6,10 @@
 
 -   Expose a JSON-RPC client, using the wallet's current JSON-RPC server.
 
+### (Breaking) Changed
+
+-   Updated API of sendTransaction and signMessage to require the account address.
+
 ## 0.1.0
 
 -   Initialized from the old browser-wallet-api-types package.